Fixed skinning
Fixes http://code.google.com/p/androido3d/issues/detail?id=2

FCollada's jointIndex is signed and has a special meaning when its
value is (-1), which O3D didn't handle. As per Collada specs,
"an index of -1 into the array of joints refers to the bind shape."

O3D maps it to the unsigned property Skin::Influence::matrix_index,
which becomes 0xffffffff in this case, and is then used as the
higher pose matrix index in various calculations, leading to a
mishandling of skinned models.

This new patch just create the joint explicitly and add it to the
list, in the Collada loader, so that O3D never sees any negative
index.

Change-Id: I69ab10da11e759ebca68d29774ab25adcc3c1550