Fixes a few floating-point related problems
This fixes an aliasing problem in bitmap.cc where uint8* was
being reinterpret as a const float* by replacing the uint*
variable with a const float* one.

This also fixes ieee-754 comparisons in various places where
the authors apparently thought floating point could be tested
for equality (they can't, as ((1.f/41.f)*41.f == 1.f) would
return false, for instance). I replaced such == and != tests
with a call to an inlined fuzzy equality testing function.

This change is needed for O3D core to build with -Wfloat-equal
and -Werror.

Change-Id: Ic1f7260538f286e8ef4132be329b569778f97aae