Return null when a model fails to load
Current Scene::LoadScene() and Scene::LoadBinaryScene()
return an error scene if the model failed to load
properly. It means the client applications have no mean
to know if there were an error or not.

This patch makes the two aforementioned methods return
NULL when the desired model could not be loaded, and
provide a separate Scene::DummyScene() that returns the
error scene if the application wants to display it.

Change-Id: I8d9585088dce1bd3bdb77e70ef44f1044e841828